feat: add  swie
