+++ packages/web3/src/ethereum/demos/siwe/index.tsx
@@ -1,4 +1,3 @@
-import { ConnectButton, Connector } from '@ant-design/web3';
 import {
   Mainnet,
   MetaMask,
@@ -8,8 +7,10 @@ import {
   WalletConnect,
 } from '@ant-design/web3-wagmi';
 import { QueryClient } from '@tanstack/react-query';
+import { createSiweMessage } from 'viem/siwe';
 import { http } from 'wagmi';
 
+import { getNonce, verifyMessage } from './mock-api';
 import SignBtn from './sign-btn';
 
 const queryClient = new QueryClient();
@@ -17,6 +18,11 @@ const queryClient = new QueryClient();
 const App: React.FC = () => {
   return (
     <WagmiWeb3ConfigProvider
+      swie={{
+        getNonce,
+        createMessage: (props) => createSiweMessage({ ...props, statement: 'Ant Design Web3' }),
+        verifyMessage: verifyMessage,
+      }}
       eip6963={{
         autoAddInjectedWallets: true,
       }}
